+++ pkg/vmd/camera_frame.go
@@ -12,7 +12,7 @@ type CameraFrame struct {
 	Curves           *CameraCurves    // 補間曲線
 }
 
-func NewCameraFrame(index int) *CameraFrame {
+func NewCameraFrame(index float32) *CameraFrame {
 	return &CameraFrame{
 		BaseFrame:        NewVmdBaseFrame(index),
 		Position:         mmath.NewMVec3(),
